Fixed Triggers, Added Servo Testing, and Updated OpMode Registry
I have fixed the trigger toggle issue with the addition of some new
booleans and if/else blocks. A new TeleOp file, ServoTester, was
created for the purpose of determining the correct position of servos.
I also updated the OpMode Register and edited some values in our
Forwards Autonomous file.