Autonomous to Diamond Shape
We worked on the autonomous and tried to make it so that the robot
would go straight, turn right, go straight again, turn right again,
then go straight. It worked, and after trying out different values and
time for each part, we managed to get it very close to stopping in the
diamond shape.