Color Sensor and Edits
ColorPressBlue is a working test program for the color sensor. Place
the robot in front of the beacon, randomize the colors, run the
program, and the robot correctly identifies red from blue.

There are some value changes in Grappling Hook as well as
AutonomousBridget1. A joystick deadzone had also been implemented in
the tele